+++ apps/website/app/(home)/blog/posts/EXAMPLE.md
