+++ apps/website/tsconfig.json
@@ -1,6 +1,9 @@
 {
   "extends": "@repo/typescript-config/nextjs.json",
   "compilerOptions": {
+    "paths": {
+      "~/*": ["./app/*"]
+    },
     "plugins": [
       {
         "name": "next"
